Created authorization middleware

+++ routes.php
@@ -1,15 +1,20 @@
 <?php
+
+// Listings
 $router->get('/', 'HomeController@index');
 $router->get('/listings', 'ListingController@index');
-$router->post('/listings', 'ListingController@store');
-$router->get('/listings/create', 'ListingController@create');
-$router->get('/listings/{id}/edit', 'ListingController@edit');
-$router->put('/listings/{id}', 'ListingController@update');
+$router->get('/listings/{id}/edit', 'ListingController@edit', ['auth']);
 $router->get('/listings/{id}', 'ListingController@show');
-$router->delete('/listings/{id}', 'ListingController@destroy');
 
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
-$router->post('/auth/register', 'UserController@store');
-$router->post('/auth/logout', 'UserController@logout');
-$router->post('/auth/login', 'UserController@authenticate');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
+$router->post('/listings', 'ListingController@store', ['auth']);
+$router->put('/listings/{id}', 'ListingController@update', ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
+
+// Users
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->get('/auth/login', 'UserController@login', ['guest'] );
+
+$router->post('/auth/register', 'UserController@store', ['guest']);
+$router->post('/auth/logout', 'UserController@logout', ['auth']);
+$router->post('/auth/login', 'UserController@authenticate', ['guest']);
